Ensure script not running concurrently.

+++ crawler/src/main/bash/pull-remote-files.sh
@@ -1,11 +1,16 @@
 #!/usr/bin/bash
+if [ -f /tmp/pull-remote-files ]; then
+    echo Script already running elsewhere.
+    exit
+fi
+touch /tmp/pull-remote-files
+echo "setex compressor-lock 3600 1" | redis-cli
 fileList=/tmp/$(openssl rand -hex 30)
 checksums=/tmp/$(openssl rand -hex 30)
 cat /var/newsreduce/network | while read host; do
     if [ $host = newsreduce.org ]; then continue; fi
     if [[ ! "$host" =~ newsreduce ]]; then continue; fi
     echo $host
-    echo "set compressor-lock 1" | redis-cli -h $host
     (rsync -alrpgoDuzhtv newsreduce@$host:/var/newsreduce/tmp/ /var/newsreduce/tmp/\
         | awk -F/ '$1=="resource"&&$2~/^[0-9]+$/&&$0~/[^\/]$/{print}'\
         > $fileList)
@@ -20,5 +25,6 @@ cat /var/newsreduce/network | while read host; do
     fi
     rm $fileList
     rm $checksums
-    echo "del compressor-lock" | redis-cli -h $host
 done
+echo "del compressor-lock" | redis-cli
+rm /tmp/pull-remote-files
